refactor(coroutines): split sample helper types into their own files
One top-level declaration for each file, per the single-responsibility
rule in the agent guide. The sample scaffolding becomes a Kit enum, a
KitPrompt object, and two private functions.

Also aligns lastOptions with lastEvent in the recording provider. Both
now fail immediately when no callback is recorded.
