+++ modules/coroutines/src/samples/kotlin/io/github/lmliam/kotventure/coroutines/prompt/Kit.kt
@@ -1,0 +1,6 @@
+package io.github.lmliam.kotventure.coroutines.prompt
+
+internal enum class Kit {
+    ARCHER,
+    MAGE,
+}
